refactor: remove DeckSection and LoginSection components from DeckManager, redirecting to login page for improved navigation
